refactor(jobs): rename demo modules to business names
Remove demo-jobs and schedule-demo naming so task modules use normal business terms across code, APIs, tests, and docs.
