+++ README.md
@@ -122,7 +122,7 @@
 - [ ] session 管理
 - [ ] 日志系统实现
 - [ ] 文件上传功能
-- [x] 定时任务系统（@nestjs/schedule 示例 + BullMQ 任务中心）
+- [x] 定时任务系统（@nestjs/schedule 轻量定时任务 + BullMQ 任务中心）
 - [ ] 单元测试与 E2E 测试
 - [ ] Docker 容器化部署
 - [ ] 国际化支持，统一管理响应信息，并根据用户语言返回不同的多语言
