+++ apps/server/src/modules/background-tasks/background-tasks.module.ts
@@ -1,12 +1,12 @@
 import { JobsModule } from '@/shared/jobs/jobs.module';
 import { Module } from '@nestjs/common';
-import { DemoJobsController } from './demo-jobs.controller';
+import { BackgroundTasksController } from './background-tasks.controller';
 import { ExportReportHandler } from './handlers/export-report.handler';
 import { FlakyRetryHandler } from './handlers/flaky-retry.handler';
 
 @Module({
   imports: [JobsModule],
-  controllers: [DemoJobsController],
+  controllers: [BackgroundTasksController],
   providers: [ExportReportHandler, FlakyRetryHandler],
 })
-export class DemoJobsModule {}
+export class BackgroundTasksModule {}
