+++ apps/server/src/modules/index.ts
@@ -1,11 +1,11 @@
 import { AuthModule } from './auth/auth.module';
 import { BenchmarkModule } from './benchmark/benchmark.module';
 import { CatsModule } from './cats/cats.module';
-import { DemoJobsModule } from './demo-jobs/demo-jobs.module';
+import { BackgroundTasksModule } from './background-tasks/background-tasks.module';
 import { GroupsModule } from './groups/groups.module';
 import { PermissionsModule } from './permissions/permissions.module';
 import { RolesModule } from './roles/roles.module';
-import { ScheduleDemoModule } from './schedule-demo/schedule-demo.module';
+import { ScheduledTasksModule } from './scheduled-tasks/scheduled-tasks.module';
 import { SocketModule } from './socket/socket.module';
 import { UsersModule } from './users/users.module';
 
@@ -13,11 +13,11 @@ export const modules = [
   AuthModule,
   BenchmarkModule,
   CatsModule,
-  DemoJobsModule,
+  BackgroundTasksModule,
   GroupsModule,
   PermissionsModule,
   RolesModule,
-  ScheduleDemoModule,
+  ScheduledTasksModule,
   SocketModule,
   UsersModule,
 ];
