+++ apps/server/src/modules/scheduled-tasks/heartbeat.scheduler.ts
@@ -2,7 +2,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 
 /**
- * 轻量进程内定时任务示例。
+ * 轻量进程内定时任务。
  * 仅用于对比 @nestjs/schedule 与 BullMQ 任务系统，不写入 job_runs。
  */
 @Injectable()
@@ -12,7 +12,7 @@ export class HeartbeatScheduler {
   @Cron(CronExpression.EVERY_MINUTE)
   handleHeartbeat() {
     this.logger.log(
-      `[schedule-demo] heartbeat pid=${process.pid} at=${new Date().toISOString()}`,
+      `[scheduled-tasks] heartbeat pid=${process.pid} at=${new Date().toISOString()}`,
     );
   }
 }
